Remove margins for page notices
They were introduced in 28e2076d6aadb8609bf76b93d4e4cfb49dbb4e1d to
prevent the notices from colliding into the page content and/or each
other, though this can no longer be reproduced on Vector, fallback, and
apioutput. Since it was also introduced 11 years ago (as of writing),
it is likely safe to deem it vestigial, and thus safe to remove.

Change-Id: Ibeb64dde952ab6280d522077a0367dcf71b086d3