Fix table syntax no longer working in page notices
Parse the notice wikitext externally and then use
recursiveTagParseFully to strip an extra <p> wrapping
the content, if present.
This allows wikitext tables and similar syntax like lists to be
interpreted correctly again.
Use Html::rawElement instead of template strings.
Also add a test to verify that tables are parsed in notices.

Bug: T392226
Change-Id: I9f522ff5de06c5d2a2b79791db607ba9c8890d50